+++ includes/mb/class.lmsensors.inc.php
@@ -28,7 +28,7 @@ class LMSensors extends Sensors
     {
         parent::__construct();
         $lines = "";
-        if ((PSI_OS == 'Linux') && !defined('PSI_EMU_HOSTNAME')) switch (defined('PSI_SENSOR_LMSENSORS_ACCESS')?strtolower(PSI_SENSOR_LMSENSORS_ACCESS):'command') {
+        if ((PSI_OS == 'Linux') && (!defined('PSI_EMU_HOSTNAME') || defined('PSI_EMU_PORT'))) switch (defined('PSI_SENSOR_LMSENSORS_ACCESS')?strtolower(PSI_SENSOR_LMSENSORS_ACCESS):'command') {
         case 'command':
             CommonFunctions::executeProgram("sensors", "", $lines);
             break;