+++ plugins/ps/class.ps.inc.php
@@ -38,7 +38,7 @@ class PS extends PSI_Plugin
         $buffer = "";
         switch (strtolower(PSI_PLUGIN_PS_ACCESS)) {
         case 'command':
-            if ((PSI_OS == 'WINNT') || defined('PSI_EMU_HOSTNAME')) {
+            if ((PSI_OS == 'WINNT') || (defined('PSI_EMU_HOSTNAME') && !defined('PSI_EMU_PORT'))) {
                 try {
                     $os_wmi = WINNT::_get_Win32_OperatingSystem();
                     $memtotal = 0;
@@ -80,7 +80,7 @@ class PS extends PSI_Plugin
                     }
                 } catch (Exception $e) {
                 }
-            } else {
+            } elseif (!defined('PSI_EMU_PORT')) {
                 CommonFunctions::executeProgram("ps", "axo pid,ppid,pmem,pcpu,args", $buffer, PSI_DEBUG);
                 if (((PSI_OS == 'Linux') || (PSI_OS == 'Android')) && (!preg_match("/^[^\n]+\n\s*\d+\s+\d+\s+[\d\.]+\s+[\d\.]+\s+.+/", $buffer))) { //alternative method if no data
                     if (CommonFunctions::rfts('/proc/meminfo', $mbuf)) {